Screen/Wayland: Implement SetMouse() mouse cursor positioning.
On Wayland servers which support the wp_pointer_warp_v1 protocol, use it
to implement PsychPositionCursor() aka SetMouse(). This allows to place
the cursor at a (x,y) position relative to the wl_surface / onscreen window
with the current pointer input focus.

This also implements coordinate remapping to deal with Retina displays
in Retina compat or full Retina mode.

Absolute cursor positioning on the desktop is not supported.
Positioning for non-focused surfaces, or outside surface bounds is not
supported.

-> So far tested with the Jay v1.11.0 Wayland compositor, the only
   compositor with support for this protocol as of August 2025.

When we are at it, pimp up MouseTraceDemo2.m for testing on Retina
displays with all bells and whistles - request full Retina resolution.

Other MouseTraceDemo's can serve for the standard "not Retina aware"
test case.