PsychVulkanCore/Linux: Now allow use of Mesa 25.3+ radv for deep color.
Mesa radv for AMD gpu's was so far marked as not supporting deep color
modes, because it didn't. Starting with Mesa 25.3 it will support deep
color, so allow probing of that driver for colorPrecision > 0 modes.

All Mesa Vulkan drivers will support deep color and HDR, starting with
Mesa 25.3, a great step forward.