Screen/Linux/Wayland: Export wl_display in winfo = Screen('GetWindowInfo', win).
winfo.SysWindowInteropHandle now contains a pointer to the internally
used wl_display object, so we can import it into PsychVulkan and thereby
PsychVulkanCore() to allow use of our Vulkan backend under Wayland, with
the wl_surface backing our onscreen window. A pointer to the wl_surface
is returned in winfo.SysWindowHandle.