PsychVulkanCore: Simplify PsychIsVulkanGPUSuitable(). Prep scRGB HDR on Wayland.
This is half-prep for using scRGB color space on Linux + Wayland for HDR.
Using scRGB is a tad more efficient/faster, but currently only supported
on KDE KWin 6.4+, not on GNOME. The current commit starts, but not completes
refactoring to detect if scRGB is supported or not. So leave the scRGB
request code half-disabled here.