Screen/PsychVulkan/Wayland: Implement backup timestamping for Vulkan+Wayland.
When using the Vulkan display backend under Wayland, and the Vulkan driver
does not support proper timing/timestamping, provide a fallback implementation
for stimulus onset timestamping which piggybacks on Screen's Wayland
timestamping: As Wayland timestamping can timestamp present completion of
a wl_surface regardless what created the surface, renders to it, or triggers
the surface.commit, and as Screen EGL/OpenGL WSI and Vulkan/WSI share the
same wl_surface for an onscreen window, lets (ab)use Screen's timestamping
for use under Vulkan as well.

The proper way to do it is to implement the Vulkan timestamping extensions
in the Vulkan drivers, but this should do as a stop-gap until that's there.