PsychVulkanCore: Skip initial PsychSetHDRMetaData() during startup when not needed.
On a conformant Vulkan driver, it should not ever be needed. Therefore only set it if native display HDR static
metadata is available, as queried from Vulkan driver or OS, or if it is needed to workaround bugs in the Vulkan
ICD. Currently the only driver known to require this to get HDR mode and HDR signalling going is AMD's
proprietary AMDVLK-Pro driver and their now deprecated AMDVLK open-source driver for Linux. For these
AMD provided ICD's, we call the function always, but on these drivers we should get valid native display HDR
metadata, so it should be fine.