Windows: Mark Windows 10 and Windows 11 as supported, all older versions as unsupported.
This was quite outdated and slipped through the cracks.

Stats as of now: ~74% are on Windows 11, ~26% on Windows 10. Then about 1 on Windows 8.x and 13 remain
on Windows 7 and 1 on Windows Vista. This for PTB 3.0.20 and later on Windows.

However, only 1 paying user left on the long unsupported Windows 7, none on Vista or 8.0 or 8.1 or any other
unsupported version. So clearly, pulling the plug completely on pre-Windows 10 in our codebase would not be
any problem.