Screen, PsychVulkan/OSX: Don't delete old finalizedFBO backing texture after assigning new one!
The idea was that once a new external Vulkan provided backing texture is assigned to the finalizedFBO, the old
now unused one should be deleted. However, this is only ok if drawBufferFBO and finalizedFBO weren't shared
in the first place (aka zero-copy redirected output buffer mode), or they were shared and are still shared after
a call to Screen('Hookfunction', win, 'SetDisplayBufferTextures', ...) -> PsychSetPipelineExportTexture() ->
PsychUnshareFinalizedFBOIfNeeded(). In this case, both drawBufferFBO and finalizedFBO will share the new
assigned backing texture, and the old backing texture will be unused and can be deleted to free up memory.

If drawBufferFBO and finalizedFBO were initially shared (zero-copy mode), but get unshared during Vulkan
interop texture setup, then the finalizedFBO will have the new interop texture assigned, but the drawBufferFBO
still uses the *old* texture which used to back both FBO's. Ergo we must not delete the old texture, as it would
pull the rug on the drawBufferFBO -> Framebuffer incomplete -> Rendering failure with lots of error messages and
a black screen.

Remove the unconditional glDeleteTextures() call from PsychVulkan.m to fix this. We add more suitable texture
release code into Screen() itself that will only delete the old texture when it is considered safe, and request its use from PsychVulkan.m.