Screen/Linux/Wayland: Add various TODO Wayland client protocol extensions.
These are the interface definition files, generated by running
wayland-scanner over suitable wayland-protocol specs, for Wayland protocol
extensions that are candidates for future implementation in Psychtoolbox
Wayland backend. The extensions are not yet implemented, implementation has
not yet started, and a few of them may not get implemented. But at the
moment, after surveying what's there, these are good potential candidates
for implementation in the near future. I've done the work of generating
these wrappers, so lets not waste it, and merge this now, as prep work,
and as a todo list and reminder.