PsychVulkanCore/OSX: Switch MoltenVK from GStreamer's to official upstream.
We still link statically against libMoltenVK.a, but get it and the headers from a
MoltenVK package from MoltenVK's official Releases page and installed side-by-side
to the Psychtoolbox-3 folder, ie. inside the same parent folder.

This gives us a much more recent, and hopefully improved MoltenVK version. Atm.
Version 1.4.1 as compared to the older 1.2 something. Some bugs were fixed and
new workarounds for Apple macOS Metal bugs implemented, some problems
remain, and are likely other not-worked-around or not-work-aroundable macOS
bugs. Especially macOS Tahoe as of macOS 26.1 is a performance trainwreck.

-> Note that building against the latest MoltenVK 1.4.1 will raise the minimum
supported macOS version to macOS 11, not only on Apple Silicon, but also on
Intel Macs!