PsychVulkanCore mex for Octave and Matlab on macOS ARM rebuilt.
This on the experimental macOS 26.1 Tahoe build system, XCode 26.1.1.

-> Matlab R2025b, Octave 10.3, MoltenVK 1.4.1 + kleinerm present-wait fix.

Something is not quite right, first run often gives vkWaitForPresentKHR() timeouts,
but successive runs then magically fix it and it works.

While the present wait fix works, and makes timing as optimal as possible, the
macOS 26.1 Tahoe performance bug persists, suggesting that the compositors
cut-off / latch deadline for a frame is literally at the beginning or very close to the
beginning of the frame, making it almost impossible for an app to latch a frame
update in time for being accepted, and thereby almost impossible to ever present
with less than 1 frame latency, causing a framerate cut to maxfps = refreshHz / 2.

Test rebuild, just as reference for further investigation.